Split test_jira_inspector from test_jira_config_loader
